pyproject.toml
[tool.poetry]
name = "netman"
version = "0.1.0"
description = "A web-based front end to manage Netgear switches"
authors = ["Bob King"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.3"
requests = "^2.32.2"
pyyaml = "^6.0.1"
flask-login = "^0.6.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
